Update api and auth

+++ src/controllers/ProductController.js
@@ -1,55 +1,104 @@
 import Product from "../models/Product.js";
+import asyncHandler from "express-async-handler";
+import User from "../models/User.js";
 
-export const getProduct = async (req, res, next) => {
+export const getProduct = asyncHandler(async (req, res, next) => {
     try {
         const products = await Product.find();
-        res.json(products);
+        res.status(200).json(products);
     } catch (err) {
         next(err);
     }
-}
+})
 
-export const getProductID = async (req, res, next) => {
+export const getProductCreatedByUser = asyncHandler(async (req, res, next) => {
     try {
-        const product = await Product.findById(req.params.id);
+        const product = await Product.find({ user: req.user.id });
         res.json(product);
     } catch (err) {
         next(err);
     }
-}
+})
 
-export const createProduct = async (req, res, next) => {
+export const createProduct = asyncHandler(async (req, res, next) => {
     try {
-        const product = new Product(req.body);
+        const product = new Product({
+            user: req.user.id,
+            name: req.body.name,
+            description: req.body.description,
+            price: req.body.price,
+            imgUrl: req.body.imgUrl,
+            quantity: req.body.quantity,
+            isAvailable: req.body.isAvailable,
+        });
         const createdProduct = await product.save();
         res.json(createdProduct);
     } catch (err) {
         next(err);
     }
-}
+})
 
-export const updatedProduct = async (req, res, next) => {
+export const updatedProduct = asyncHandler(async (req, res, next) => {
     try {
+        const findProduct = await Product.findById(req.params.id);
+        if (!findProduct) {
+            res.status(404);
+            throw new Error("Product not found");
+        }
+
+        const user = await User.findById(req.user.id);
+        // check for user
+        if (!user) {
+            res.status(404);
+            throw new Error("User not found");
+        }
+
+        // check if user is the owner of the product
+        if (findProduct.user.toString() !== req.user.id) {
+            res.status(401);
+            throw new Error("User not authorized");
+        }
+
         const product = await Product.findByIdAndUpdate(req.params.id);
+
         product.name = req.body.name;
         product.description = req.body.description;
         product.price = req.body.price;
         product.imgUrl = req.body.imgUrl;
         product.quantity = req.body.quantity;
-        product.isFeatured = req.body.isFeatured;
+        product.isAvailable = req.body.isAvailable;
+        
         const updatedProduct = await product.save();
         res.json(updatedProduct);
     } catch (err) {
         next(err);
     }
-}
+})
 
-export const deleteProduct = async (req, res, next) => {
+export const deleteProduct = asyncHandler(async (req, res, next) => {
     try {
+        const findProduct = await Product.findById(req.params.id);
+        if (!findProduct) {
+            res.status(404);
+            throw new Error("Product not found");
+        }
+
+        const user = await User.findById(req.user.id);
+        // check for user
+        if (!user) {
+            res.status(404);
+            throw new Error("User not found");
+        }
+
+        // check if user is the owner of the product
+        if (findProduct.user.toString() !== req.user.id) {
+            res.status(401);
+            throw new Error("User not authorized");
+        }
         const product = await Product.findByIdAndDelete(req.params.id);
         res.json(product);
     } catch (err) {
         next(err);
     }
-}
+})
